+++ familyhelper-finance-impl/src/main/java/com/dwarfeng/familyhelper/finance/impl/bean/entity/HibernateFundChange.java
@@ -5,6 +5,7 @@ import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Optional;
 
@@ -13,7 +14,7 @@ import java.util.Optional;
 @Table(name = "tbl_fund_change")
 public class HibernateFundChange implements Bean {
 
-    private static final long serialVersionUID = -5335836666782885937L;
+    private static final long serialVersionUID = 1203465328971020251L;
 
     // -----------------------------------------------------------主键-----------------------------------------------------------
     @Id
@@ -24,10 +25,10 @@ public class HibernateFundChange implements Bean {
     @Column(name = "account_book_id")
     private Long accountBookLongId;
 
-    @Column(name = "bank_card_id")
-    private Long bankCardLongId;
-
     // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    @Column(name = "delta", nullable = false)
+    private BigDecimal delta;
+
     @Column(name = "change_type", length = Constraints.LENGTH_TYPE)
     private String changeType;
 
@@ -44,12 +45,6 @@ public class HibernateFundChange implements Bean {
             @JoinColumn(name = "account_book_id", referencedColumnName = "id", insertable = false, updatable = false), //
     })
     private HibernateAccountBook accountBook;
-
-    @ManyToOne(targetEntity = HibernateBankCard.class)
-    @JoinColumns({ //
-            @JoinColumn(name = "bank_card_id", referencedColumnName = "id", insertable = false, updatable = false), //
-    })
-    private HibernateBankCard bankCard;
 
     public HibernateFundChange() {
     }
@@ -71,14 +66,6 @@ public class HibernateFundChange implements Bean {
         this.accountBookLongId = Optional.ofNullable(idKey).map(HibernateLongIdKey::getLongId).orElse(null);
     }
 
-    public HibernateLongIdKey getBankCardKey() {
-        return Optional.ofNullable(bankCardLongId).map(HibernateLongIdKey::new).orElse(null);
-    }
-
-    public void setBankCardKey(HibernateLongIdKey idKey) {
-        this.bankCardLongId = Optional.ofNullable(idKey).map(HibernateLongIdKey::getLongId).orElse(null);
-    }
-
     // -----------------------------------------------------------常规属性区-----------------------------------------------------------
     public Long getLongId() {
         return longId;
@@ -96,12 +83,12 @@ public class HibernateFundChange implements Bean {
         this.accountBookLongId = accountBookLongId;
     }
 
-    public Long getBankCardLongId() {
-        return bankCardLongId;
+    public BigDecimal getDelta() {
+        return delta;
     }
 
-    public void setBankCardLongId(Long bankCardLongId) {
-        this.bankCardLongId = bankCardLongId;
+    public void setDelta(BigDecimal delta) {
+        this.delta = delta;
     }
 
     public String getChangeType() {
@@ -136,24 +123,15 @@ public class HibernateFundChange implements Bean {
         this.accountBook = accountBook;
     }
 
-    public HibernateBankCard getBankCard() {
-        return bankCard;
-    }
-
-    public void setBankCard(HibernateBankCard bankCard) {
-        this.bankCard = bankCard;
-    }
-
     @Override
     public String toString() {
         return getClass().getSimpleName() + "(" +
                 "longId = " + longId + ", " +
                 "accountBookLongId = " + accountBookLongId + ", " +
-                "bankCardLongId = " + bankCardLongId + ", " +
+                "delta = " + delta + ", " +
                 "changeType = " + changeType + ", " +
                 "happenedDate = " + happenedDate + ", " +
                 "remark = " + remark + ", " +
-                "accountBook = " + accountBook + ", " +
-                "bankCard = " + bankCard + ")";
+                "accountBook = " + accountBook + ")";
     }
 }