+++ familyhelper-finance-impl/src/test/java/com/dwarfeng/familyhelper/finance/impl/service/FundChangeMaintainServiceImplTest.java
@@ -1,10 +1,8 @@
 package com.dwarfeng.familyhelper.finance.impl.service;
 
 import com.dwarfeng.familyhelper.finance.stack.bean.entity.AccountBook;
-import com.dwarfeng.familyhelper.finance.stack.bean.entity.BankCard;
 import com.dwarfeng.familyhelper.finance.stack.bean.entity.FundChange;
 import com.dwarfeng.familyhelper.finance.stack.service.AccountBookMaintainService;
-import com.dwarfeng.familyhelper.finance.stack.service.BankCardMaintainService;
 import com.dwarfeng.familyhelper.finance.stack.service.FundChangeMaintainService;
 import org.apache.commons.beanutils.BeanUtils;
 import org.junit.After;
@@ -30,12 +28,9 @@ public class FundChangeMaintainServiceImplTest {
     private FundChangeMaintainService fundChangeMaintainService;
     @Autowired
     private AccountBookMaintainService accountBookMaintainService;
-    @Autowired
-    private BankCardMaintainService bankCardMaintainService;
 
     private List<FundChange> fundChanges;
     private AccountBook accountBook;
-    private BankCard bankCard;
 
     @Before
     public void setUp() {
@@ -44,7 +39,7 @@ public class FundChangeMaintainServiceImplTest {
             FundChange fundChange = new FundChange(
                     null,
                     null,
-                    null,
+                    BigDecimal.ZERO,
                     "change_type",
                     new Date(),
                     "remark"
@@ -58,25 +53,12 @@ public class FundChangeMaintainServiceImplTest {
                 BigDecimal.ZERO,
                 "remark"
         );
-        bankCard = new BankCard(
-                null,
-                null,
-                "name",
-                "card_type",
-                new Date(),
-                BigDecimal.ZERO,
-                true,
-                new Date(),
-                BigDecimal.ZERO,
-                "remark"
-        );
     }
 
     @After
     public void tearDown() {
         fundChanges.clear();
         accountBook = null;
-        bankCard = null;
     }
 
     @Test
@@ -115,32 +97,6 @@ public class FundChangeMaintainServiceImplTest {
             ).getCount());
         } finally {
             accountBookMaintainService.deleteIfExists(accountBook.getKey());
-            for (FundChange fundChange : fundChanges) {
-                fundChangeMaintainService.deleteIfExists(fundChange.getKey());
-            }
-        }
-    }
-
-    @Test
-    public void testForBankCardCascade() throws Exception {
-        try {
-            bankCard.setKey(bankCardMaintainService.insertOrUpdate(bankCard));
-            for (FundChange fundChange : fundChanges) {
-                fundChange.setBankCardKey(bankCard.getKey());
-                fundChange.setKey(fundChangeMaintainService.insert(fundChange));
-            }
-
-            assertEquals(fundChanges.size(), fundChangeMaintainService.lookup(
-                    FundChangeMaintainService.CHILD_FOR_BANK_CARD, new Object[]{bankCard.getKey()}
-            ).getCount());
-
-            bankCardMaintainService.deleteIfExists(bankCard.getKey());
-
-            assertEquals(0, fundChangeMaintainService.lookup(
-                    FundChangeMaintainService.CHILD_FOR_BANK_CARD, new Object[]{bankCard.getKey()}
-            ).getCount());
-        } finally {
-            bankCardMaintainService.deleteIfExists(bankCard.getKey());
             for (FundChange fundChange : fundChanges) {
                 fundChangeMaintainService.deleteIfExists(fundChange.getKey());
             }